Correct the fix to previews not being cleared on rebuilding them
That fix was in commit d9c5fefc546857c483924f635fadf36fd4dd1041.

It was mistaken because the clearing of previews should be conditional
on the rebuild scope plugin setting being "All".

+++ root/inc/plugins/linktools.php
@@ -3748,8 +3748,10 @@ function lkt_hookin__admin_tools_recount_rebuild() {
 			if($mybb->input['page'] == 1) {
 				// Log admin action
 				log_admin_action($lang->lkt_admin_log_rebuild_linkpreviews);
-				$db->update_query('urls', array('got_preview' => 0));
-				$db->delete_query('url_previews');
+				if ($mybb->settings[C_LKT.'_link_preview_rebuild_scope'] == 'all') {
+					$db->update_query('urls', array('got_preview' => 0));
+					$db->delete_query('url_previews');
+				}
 			}
 
 			if (!$mybb->get_input('lkt_linkpreviews_per_page', MyBB::INPUT_INT)) {
